+++ application/core/MY_Controller.php
@@ -25,6 +25,10 @@ class Application extends CI_Controller
 		$this->data = array ();
 		$this->data['pagetitle'] = 'Sam & Ella\'s';
 		$this->data['ci_version'] = (ENVIRONMENT === 'development') ? 'CodeIgniter Version <strong>'.CI_VERSION.'</strong>' : '';
+
+		// get the user role
+		$this->data['userrole'] = $this->session->userdata('userrole');
+		if ($this->data['userrole'] == NULL) $this->data['userrole'] = '?';
 	}
 
 	/**
@@ -36,6 +40,11 @@ class Application extends CI_Controller
 	    $this->data['navbar'] = $this->parser->parse('navbar', $this->data, true);
 	    if (!isset($this->data['content']))
 	        $this->data['content'] = $this->parser->parse($this->data['pagebody'], $this->data, true);
+            
+            $this->data['caboose_styles'] = $this->caboose->styles();
+            $this->data['caboose_scripts'] = $this->caboose->scripts();
+            $this->data['caboose_trailings'] = $this->caboose->trailings();
+        
 	    $this->parser->parse($template, $this->data);
 	}