+++ data/starter-data.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Host: localhost
--- Generation Time: Oct 16, 2016 at 11:01 PM
+-- Generation Time: Oct 19, 2016 at 12:10 AM
 -- Server version: 5.7.13
 -- PHP Version: 7.0.8
 
@@ -26,7 +26,8 @@ SET time_zone = "+00:00";
 -- Table structure for table `categories`
 --
 
-CREATE TABLE `categories` (
+DROP TABLE IF EXISTS `Categories`;
+CREATE TABLE `Categories` (
   `id` varchar(1) NOT NULL,
   `name` varchar(64) NOT NULL,
   `description` text NOT NULL,
@@ -37,7 +38,7 @@ CREATE TABLE `categories` (
 -- Dumping data for table `categories`
 --
 
-INSERT INTO `categories` (`id`, `name`, `description`, `image`) VALUES
+INSERT INTO `Categories` (`id`, `name`, `description`, `image`) VALUES
 ('d', 'Drinks', 'Purees made from the finest of Venusian insects, government-inspected.', 'catd.png'),
 ('m', 'Mains', 'Made from only the finest ingredients found deep in the Martian jungle, and harvested or slaughtered by academy-trained druids, we bring you 45 day aged premium "meat".', 'catm.png'),
 ('s', 'Sides', 'Perfect accompaniments to our mains, these side dish pairings have been exclsisvely formulated by Ben & Jerry.', 'cats.png');
@@ -45,10 +46,24 @@ INSERT INTO `categories` (`id`, `name`, `description`, `image`) VALUES
 -- --------------------------------------------------------
 
 --
+-- Table structure for table `ci_sessions`
+--
+
+CREATE TABLE `ci_sessions` (
+  `id` varchar(128) NOT NULL,
+  `ip_address` varchar(45) NOT NULL,
+  `timestamp` int(10) UNSIGNED NOT NULL DEFAULT '0',
+  `data` blob NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+-- --------------------------------------------------------
+
+--
 -- Table structure for table `menu`
 --
 
-CREATE TABLE `menu` (
+DROP TABLE IF EXISTS `Menu`;
+CREATE TABLE `Menu` (
   `id` int(11) NOT NULL,
   `name` varchar(32) NOT NULL,
   `description` varchar(256) NOT NULL,
@@ -61,7 +76,7 @@ CREATE TABLE `menu` (
 -- Dumping data for table `menu`
 --
 
-INSERT INTO `menu` (`id`, `name`, `description`, `price`, `picture`, `category`) VALUES
+INSERT INTO `Menu` (`id`, `name`, `description`, `price`, `picture`, `category`) VALUES
 (1, 'Cheese', 'Leave this raw milk, beefy and sweet cheese out for an hour before serving and pair with pear jam.', '2.95', '1.png', 's'),
 (2, 'Turkey', 'Roasted, succulent, stuffed, lovingly sliced turkey breast', '5.95', '2.png', 'm'),
 (6, 'Donut', 'Disgustingly sweet, topped with artery clogging chocolate and then sprinkled with Pixie dust', '1.25', '6.png', 's'),
@@ -77,7 +92,8 @@ INSERT INTO `menu` (`id`, `name`, `description`, `price`, `picture`, `category`)
 -- Table structure for table `orderitems`
 --
 
-CREATE TABLE `orderitems` (
+DROP TABLE IF EXISTS `Orderitems`;
+CREATE TABLE `Orderitems` (
   `order` int(11) NOT NULL,
   `item` int(11) NOT NULL,
   `quantity` int(11) NOT NULL,
@@ -90,7 +106,8 @@ CREATE TABLE `orderitems` (
 -- Table structure for table `orders`
 --
 
-CREATE TABLE `orders` (
+DROP TABLE IF EXISTS `Orders`;
+CREATE TABLE `Orders` (
   `num` int(11) NOT NULL,
   `date` datetime NOT NULL,
   `status` varchar(1) NOT NULL,
@@ -105,25 +122,31 @@ CREATE TABLE `orders` (
 --
 -- Indexes for table `categories`
 --
-ALTER TABLE `categories`
+ALTER TABLE `Categories`
   ADD PRIMARY KEY (`id`);
+
+--
+-- Indexes for table `ci_sessions`
+--
+ALTER TABLE `ci_sessions`
+  ADD KEY `ci_sessions_timestamp` (`timestamp`);
 
 --
 -- Indexes for table `menu`
 --
-ALTER TABLE `menu`
+ALTER TABLE `Menu`
   ADD PRIMARY KEY (`id`);
 
 --
 -- Indexes for table `orderitems`
 --
-ALTER TABLE `orderitems`
+ALTER TABLE `Orderitems`
   ADD PRIMARY KEY (`order`,`item`);
 
 --
 -- Indexes for table `orders`
 --
-ALTER TABLE `orders`
+ALTER TABLE `Orders`
   ADD PRIMARY KEY (`num`);
 
 /*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;